fix(ci): resolve MinGW format warnings and clang-tidy failures
- trace.cpp: cast ULONG params to unsigned int before printf's %x/%u -
  ULONG is `unsigned long` under MinGW/Windows, format-checked strictly there.
- windows_emulator.cpp: remove the dead APISETDIAG byte-scan block
  (hook_memory_execution never fires under the FEX backend - confirmed
  across 4 placement attempts this investigation - so this code never ran
  and only tripped cppcoreguidelines-avoid-c-arrays/modernize-avoid-c-arrays).
- windows_emulator.cpp: use auto instead of restating the cast's type for
  walk_ebp (modernize-use-auto).

+++ src/windows-emulator/syscalls/trace.cpp
@@ -91,8 +91,9 @@ namespace sogen
                 fprintf(stderr,
                         "[APISETWATCH] NtTraceControl function_code=0x%x base=0x%x input_buffer=0x%llx "
                         "input_buffer_length=%u output_buffer=0x%llx output_buffer_length=%u\n",
-                        function_code, base_function_code, static_cast<unsigned long long>(input_buffer), input_buffer_length,
-                        static_cast<unsigned long long>(output_buffer), output_buffer_length);
+                        static_cast<unsigned int>(function_code), static_cast<unsigned int>(base_function_code),
+                        static_cast<unsigned long long>(input_buffer), static_cast<unsigned int>(input_buffer_length),
+                        static_cast<unsigned long long>(output_buffer), static_cast<unsigned int>(output_buffer_length));
                 fflush(stderr);
             }
